requirements.txt
torchvision==0.4.1+cpu -f https://download.pytorch.org/whl/torch_stable.html
torch==1.3.1+cpu -f https://download.pytorch.org/whl/torch_stable.html
fastai
numpy
pathtools
streamlit
Pillow==6.1.0
